Confidence Probabilities

Estimate confidence probabilities ('backward') and assignment error rates ('forward') per category (genotyped/dummy) by repeatedly simulating genotype data from a reference pedigree using SimGeno, reconstruction a pedigree from this using sequoia, and counting the number of mismatches using PedCompare.

EstConf(
  Pedigree = NULL,
  LifeHistData = NULL,
  args.sim = list(nSnp = 400, SnpError = 0.001, ParMis = c(0.4, 0.4)),
  args.seq = list(Module = "ped", Err = 0.001, Tassign = 0.5, CalcLLR = FALSE),
  nSim = 10,
  nCores = 1,
  quiet = TRUE
)

Arguments

Pedigree
reference pedigree from which to simulate, dataframe with columns id-dam-sire. Additional columns are ignored.

LifeHistData
dataframe with id, sex (1=female, 2=male, 3=unknown), and birth year.

args.sim
list of arguments to pass to SimGeno, such as nSnp (number of SNPs), SnpError (genotyping error rate) and ParMis (proportion of non-genotyped parents). Set to NULL to use all default values.

args.seq
list of arguments to pass to sequoia, such as Module ('par' or 'ped'), Err (assumed genotyping error rate), and Complex. May include (part of) SeqList, a list of sequoia output (i.e. as a list-within-a-list). Set to NULL to use all default values.

nSim
number of iterations of simulate - reconstruct - compare to perform, i.e. number of simulated datasets.

nCores
number of computer cores to use. If >1, package parallel is used. Set to NULL to use all but one of the available cores, as detected by parallel::detectCores() (using all cores tends to freeze up your computer).

quiet
suppress messages. TRUE runs SimGeno and sequoia quietly, 'very' also suppresses other messages and the iteration counter when nCores=1 (there is no iteration counter when nCores>1).

Value

A list, with elements:

ConfProb

See below

PedErrors

See below

Pedigree.reference

the pedigree from which data was simulated

LifeHistData

Pedigree.inferred

a list with for each iteration the inferred pedigree based on the simulated data

SimSNPd

a list with for each iteration the IDs of the individuals simulated to have been genotyped

PedComp.fwd

Counts from the 'forward' PedCompare, from which PedErrors is calculated

RunParams

a list with the call to EstConf, as well as the default parameter values for SimGeno, and sequoia.

RunTime

sequoia runtime per simulation in seconds, as measured by system.time()['elapsed'].

Dataframe ConfProb has 7 columns:
id.cat, dam.cat, sire.cat

Category of the focal individual, dam, and sire, in the pedigree inferred based on the simulated data. Coded as G=genotyped, D=dummy, X=none

dam.conf

Probability that the dam is correct, given the categories of the assigned dam and sire (ignoring whether or not the sire is correct)

sire.conf

as dam.conf, for the sire

pair.conf

Probability that both dam and sire are correct, given their categories

N

Number of individuals per category-combination, across all nSim iterations

Array PedErrors has three dimensions:
class

FalseNeg(atives): could have been assigned but was not (individual + parent both genotyped or dummyfiable; P1only in PedCompare).

FalsePos(itives): no parent in reference pedigree, but one was assigned based on the simulated data (P2only)

Mismatch: different parents between the pedigrees

cat

Category of individual + parent, as a two-letter code where the first letter indicates the focal individual and the second the parent; G=Genotyped, D=Dummy, T=Total

parent

dam or sire

Details

The confidence probability is taken as the number of correct (matching) assignments, divided by all assignments made in the observed (inferred-from-simulated) pedigree. In contrast, the false negative & false positive assignment rates are proportions of the number of parents in the true (reference) pedigree. Each rate is calculated separatedly for dams & sires, and separately for each category (Genotyped/Dummy(fiable)/X (none)) of individual, parent and co-parent.

This function does not know which individuals in the actual Pedigree are genotyped, so the confidence probabilities need to be added to the Pedigree as shown in the example at the bottom.

A confidence of \(1\) means all assignments on simulated data were correct for that category-combination. It should be interpreted as (and perhaps modified to) \(> 1 - 1/N\), where sample size N is given in the last column of the ConfProb and PedErrors dataframes in the output. The same applies for a false negative/positive rate of \(0\) (i.e. to be interpreted as \(< 1/N\)).

Assumptions

Because the actual true pedigree is (typically) unknown, the provided reference pedigree is used as a stand-in and assumed to be the true pedigree, with unrelated founders. It is also assumed that the probability to be genotyped is equal for all parents; in each iteration, a new random set of parents (proportion set by ParMis) is mimicked to be non-genotyped. In addition, SNPs are assumed to segregate independently.

Object size

The size in Kb of the returned list can become pretty big, as each of the inferred pedigrees is included. When running EstConf many times for a range of parameter values, it may be prudent to save the required summary statistics for each run rather than the full output.

See also

Examples

# \donttest{
data(Ped_HSg5, LH_HSg5, package="sequoia")

## Example A: parentage assignment only
conf.A <- EstConf(Pedigree = Ped_HSg5, LifeHistData = LH_HSg5,
   args.sim = list(nSnp = 100, SnpError = 5e-3, ParMis=c(0.2, 0.5)),
   args.seq = list(Module="par", Err=1e-3, Tassign=0.5), nSim = 3)

#> Simulating parentage assignment only ...

#> i= 1 	 19:19:06 
#> i= 2 	 19:19:11 
#> i= 3 	 19:19:17 

# parent-pair confidence, per category:
conf.A$ConfProb

#>   id.cat dam.cat sire.cat dam.conf sire.conf pair.conf    N
#> 1      G       G        G        1    0.9991    0.9991 1074
#> 2      G       G        X        0        NA        NA  981
#> 3      G       X        G       NA    0.0000        NA  291
#> 4      G       X        X       NA        NA        NA  462

# calculate (correct) assignment rates (ignores co-parent)
1 - apply(conf.A$PedErrors, c(1,3), sum, na.rm=TRUE)

#>     parent
#> cat        dam      sire
#>   GG 0.9372347 0.9787446
#>   GD 0.0000000 0.0000000
#>   GT 0.7506371 0.4899128
#>   DG 0.0000000 0.0000000
#>   DD 0.0000000 0.0000000
#>   DT 0.0000000 0.0000000
#>   TT 0.7107639 0.4638889

## Example B: with sibship clustering, based on sequoia inferred pedigree
RealGenotypes <- SimGeno(Ped = Ped_HSg5, nSnp = 100,
                         ParMis=c(0.19,0.53), SnpError = 6e-3)
SeqOUT <- sequoia(GenoM = RealGenotypes,
                  LifeHistData = LH_HSg5,
                  Err=5e-3, Module="ped",
                  quiet=TRUE, Plot=FALSE)

conf.B <- EstConf(Pedigree = SeqOUT$Pedigree,
              LifeHistData = LH_HSg5,
               args.sim = list(nSnp = 100, SnpError = 5e-3,
                               ParMis=c(0.2, 0.5)),
              args.seq = list(Err=5e-3, Module="ped"),
              nSim = 2, nCores=2)

#> Simulating full pedigree reconstruction ...

#> Using 2 out of 6 cores

conf.B$ConfProb

#>    id.cat dam.cat sire.cat dam.conf sire.conf pair.conf   N
#> 1       G       G        G    0.989     0.980     0.979 663
#> 2       G       G        D    1.000     0.914     0.914 748
#> 3       G       G        X    1.000        NA        NA  12
#> 4       G       D        G    0.878     0.928     0.878 180
#> 5       G       D        D    0.934     0.890     0.835 182
#> 6       G       D        X    0.750        NA        NA   4
#> 7       G       X        G       NA     1.000        NA   7
#> 8       G       X        D       NA     1.000        NA   6
#> 9       G       X        X       NA        NA        NA  72
#> 10      D       G        G    1.000     1.000     1.000  34
#> 11      D       G        D    1.000     0.951     0.951  48
#> 12      D       G        X    0.000        NA        NA   7
#> 13      D       D        G    0.778     0.852     0.778  15
#> 14      D       D        D    0.754     0.877     0.754  13
#> 15      D       D        X      NaN        NA        NA   0
#> 16      D       X        G       NA     0.000        NA   1
#> 17      D       X        D       NA     0.000        NA   1
#> 18      D       X        X       NA        NA        NA  30

Ped.withConf <- getAssignCat(Pedigree = SeqOUT$Pedigree,
                             SNPd = rownames(RealGenotypes))
Ped.withConf <- merge(Ped.withConf, conf.B$ConfProb, all.x=TRUE, sort=FALSE)
Ped.withConf <- Ped.withConf[, c("id","dam","sire", "dam.conf", "sire.conf",
                                 "id.cat", "dam.cat", "sire.cat")]
head(Ped.withConf[Ped.withConf$dam.cat=="G", ])

#>        id    dam  sire dam.conf sire.conf id.cat dam.cat sire.cat
#> 33 a01002 a00014 M0019        1     0.914      G       G        D
#> 34 a01001 a00014 M0019        1     0.914      G       G        D
#> 35 b01006 a00013 M0034        1     0.914      G       G        D
#> 36 b01007 a00013 M0034        1     0.914      G       G        D
#> 37 a01008 a00013 M0034        1     0.914      G       G        D
#> 38 b01009 a00008 M0015        1     0.914      G       G        D

head(Ped.withConf[Ped.withConf$dam.cat=="D", ])

#>         id   dam  sire dam.conf sire.conf id.cat dam.cat sire.cat
#> 406 b02186 F0016 M0035    0.934      0.89      G       D        D
#> 407 b03020 F0013 M0024    0.934      0.89      G       D        D
#> 408 b02092 F0003 M0009    0.934      0.89      G       D        D
#> 409 a05058 F0023 M0011    0.934      0.89      G       D        D
#> 410 b02128 F0017 M0036    0.934      0.89      G       D        D
#> 411 b04097 F0024 M0016    0.934      0.89      G       D        D

## P(actual FS | inferred as FS) etc.
PairL <- list()
for (i in 1:length(conf.A$Pedigree.inferred)) {  # nSim
  cat(i, "\t")
  PairL[[i]] <- ComparePairs(conf.A$Pedigree.reference,
                             conf.A$Pedigree.inferred[[i]],
                             GenBack=1, patmat=TRUE, ExcludeDummies = TRUE,
                             Return="Counts")
}

#> 1 	2 	3 

# P(actual relationship (Ped1) | inferred relationship (Ped2))
PairA <- plyr::laply(PairL, function(M) sweep(M, 2, colSums(M), "/"))
PairRel.prop <- apply(PairA, 2:3, mean, na.rm=TRUE)  # mean across simulations
round(PairRel.prop, 2)

#>      Ped2
#> Ped1  M    P O FS  MHS  PHS    U    X
#>   M   1 0.00 0  0 0.00 0.00 0.00 0.00
#>   P   0 0.98 0  0 0.00 0.00 0.00 0.01
#>   FS  0 0.00 0  1 0.30 0.08 0.00 0.00
#>   MHS 0 0.00 0  0 0.69 0.00 0.00 0.00
#>   PHS 0 0.00 0  0 0.00 0.91 0.00 0.01
#>   U   0 0.02 1  0 0.01 0.01 0.99 0.98
#>   X   0 0.00 0  0 0.00 0.00 0.00 0.00

#' # or: P(inferred relationship | actual relationship)
PairA2 <- plyr::laply(PairL, function(M) sweep(M, 1, rowSums(M), "/"))
# }