preparations

Check parameter values and save as named vector.

SeqPrep(
  GenoM = NULL,
  LifeHistData = NULL,
  nAgeClasses = 1,
  MaxSibIter = 5,
  Err = 1e-04,
  MaxMismatchV = NULL,
  Tfilter = -2,
  Tassign = 0.5,
  MaxSibshipSize = 100,
  DummyPrefix = c("F", "M"),
  Complexity = "full",
  UseAge = "yes",
  FindMaybeRel = TRUE,
  CalcLLR = TRUE
)

Arguments

GenoM
matrix with genotype data, size nInd x nSnp

LifeHistData
Dataframe with 3 columns:

ID

max. 30 characters long,

Sex

1 = females, 2 = males, other numbers = unknown,

Birth Year

(or hatching year) Zero and negative numbers are interpreted as missing values.

nAgeClasses
Number of age classes (= no. rows in AgePriors)

MaxSibIter
Maximum number of iterations of sibship clustering (up to 42).

Err
Estimated genotyping error rate.

MaxMismatchV
Maximum number of loci at which (1) a duplicate sample mismatches; (2) candidate parent and offspring are allowed to be opposite homozygotes; (3) parent-parent-offspring trios can have Mendelian errors.

Tfilter
Threshold log-likelihood ratio between a proposed relationship versus unrelated, to select candidate relatives. Typically a negative value, related to the fact that unconditional likelihoods are calculated during the filtering steps. More negative values may decrease non-assignment, but will increase computational time.

Tassign
Minimum log-likelihood ratio required for acceptance of proposed relationship, relative to next most likely relationship. Higher values result in more conservative assignments.

MaxSibshipSize
Maximum number of offspring for a single individual (a generous safety margin is advised).

DummyPrefix
character vector of length 2 with prefixes for dummy dams (mothers) and sires (fathers); maximum 20 characters each.

Complexity
Either "full" (default), "simp" (no explicit consideration of inbred relationships), "mono" (monogamous breeding system), or "herm" (hermaphrodites)

FindMaybeRel
Identify pairs of non-assigned likely relatives after pedigree reconstruction. Can be time-consuming in large datasets.

CalcLLR
Calculate log-likelihood ratios for all assignments. Can be time-consuming in large datasets.

Value

A 1-row dataframe with parameter values

Details

Please do not increasing the number of SNPs or individuals beyond the numbers present in the datasets, as this may cause R to crash.